pyproject.toml
[tool.poetry]
name = "api_test_with_env"
version = "1.0"
description = "API test`s using env"
authors = ["qagurutestapi <qatestapidemo@gmail.ru>"]


[tool.poetry.dependencies]
python = "^3.10"
selene = {version = "^2.0.0-beta.8", allow-prereleases = true}
pytest = "^7.1.3"
allure-pytest = "^2.10.0"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
curlify = "^2.2.1"
pytest-voluptuous = "^1.2.0"